reorganize utility functions
