+++ odtlearn/FairTree.py
@@ -5,8 +5,8 @@ from sklearn.utils.validation import check_array, check_is_fitted, check_X_y
 
 from odtlearn.tree_classifier import TreeClassifier
 from odtlearn.utils.strongtree_formulation import FairOCT
-from odtlearn.utils.StrongTreeUtils import check_binary, check_columns_match
 from odtlearn.utils.Tree import _Tree
+from odtlearn.utils.validation import check_binary, check_columns_match
 
 
 class FairTreeClassifier(TreeClassifier):
